[update] split struct definition to library carate and rename project name
